Win32 (MSVC): Cannot use variadic token pasting GNU ext
I have read [1] that MSVC gets rid of any trailing ',' but I have not
seen evidence (*) of this so I dropped comma before the __VA_ARGS__.
This may result in the __FILE__ argument being combine with the first
variadic argument.

This change is untested (except that it compiles in the case where it was
broken) because I cannot make a debug build without debug Boost libraries
(and probably debug Qt libraries too).

There are cross-platform approaches [1] that could be used if needs be.

[1] http://stackoverflow.com/questions/5588855/standard-alternative-to-gccs-va-args-trick

* This commit assumes that what MSVC does is to insert a ',' when needed.